Normalize column structure by separating category IDs
This creates a new column family named `category_id_to_name` which
resolves category IDs (really, page IDs that happen to be categories)
to the string names of the categories. This way a list of
subcategories may be stored as a `u64[]` instead of a
`str[]`. Theoretically, this is a normalization which should lower
disk space requirements, which is the primary goal for this service.